Redesign
See https://github.com/PistonDevelopers/piston/issues/933
