--- src/event/.gitignore
@@ -1,31 +1,0 @@
-.DS_Store
-*~
-*#
-*.o
-*.so
-*.swp
-*.old
-*.bak
-*.kate-swp
-*.dylib
-*.dSYM
-*.dll
-*.rlib
-*.dummy
-*.exe
-*-test
-/target/main
-/target/test-internal
-/target/test-external
-/doc/
-/target/
-/build/
-/.rust/
-rusti.sh
-watch.sh
-/examples/**
-!/examples/*.rs
-!/examples/assets/
-!/target/assets/
-Cargo.lock
-
