--- src/event/README.md
@@ -1,20 +1,0 @@
-# event [![Build Status](https://travis-ci.org/PistonDevelopers/piston.svg)](https://travis-ci.org/PistonDevelopers/piston)[![Crates.io](https://img.shields.io/crates/v/piston.svg?style=flat-square)](https://crates.io/crates/piston)
-
-A library for flexible generic event threading
-
-This library is reexported under [Piston](https://github.com/pistondevelopers/piston)
-
-[Online docs](http://docs.piston.rs/piston/piston/event/index.html)
-
-[How to contribute](https://github.com/PistonDevelopers/piston/blob/master/CONTRIBUTING.md)
-
-### How to build
-
-Install [Rust](http://www.rust-lang.org/)
-
-`cargo build`
-
-### Dependency graph
-
-![event](./Cargo.png)
-
