+++ src/event_loop/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 
 name = "pistoncore-event_loop"
-version = "0.4.0"
+version = "0.5.0"
 authors = [
     "bvssvni <bvssvni@gmail.com>",
     "Coeuvre <coeuvre@gmail.com>",
@@ -21,6 +21,10 @@ name = "event_loop"
 path = "src/lib.rs"
 
 [dependencies.pistoncore-window]
+path = "../window"
+version = "0.4.0"
+
+[dependencies.pistoncore-input]
 path = "../window"
 version = "0.3.0"
