+++ src/input/Cargo.toml
@@ -1,7 +1,6 @@
 [package]
-
 name = "pistoncore-input"
-version = "0.2.0"
+version = "0.3.0"
 authors = ["bvssvni <bvssvni@gmail.com>"]
 keywords = ["keyboard", "mouse", "input", "piston"]
 description = "A structure for user input"
@@ -11,12 +10,11 @@ repository = "https://github.com/PistonDevelopers/piston.git"
 homepage = "https://github.com/PistonDevelopers/piston"
 
 [lib]
-
 name = "input"
 path = "src/lib.rs"
 
 [dependencies]
-
 rustc-serialize = "0.3.15"
 bitflags = "0.3.2"
 num = "0.1.25"
+piston-viewport = "0.1.0"
