+++ src/input/src/event.rs
@@ -1,17 +1,14 @@
-use input;
-
 use {
-    EventMap,
     IdleArgs,
+    Input,
     RenderArgs,
     AfterRenderArgs,
     UpdateArgs,
 };
-use Event::{ Idle, Render, AfterRender, Update, Input };
 
 /// Adds render and update events to input events
 #[derive(Clone, PartialEq, Debug)]
-pub enum Event<I = input::Input> {
+pub enum Event<I = Input> {
     /// Render graphics.
     Render(RenderArgs),
     /// After rendering and swapping buffers.
@@ -22,12 +19,4 @@ pub enum Event<I = input::Input> {
     Idle(IdleArgs),
     /// Input event.
     Input(I),
-}
-
-impl<I> EventMap<I> for Event<I> {
-    fn render(args: RenderArgs) -> Event<I> { Render(args) }
-    fn after_render(args: AfterRenderArgs) -> Event<I> { AfterRender(args) }
-    fn update(args: UpdateArgs) -> Event<I> { Update(args) }
-    fn idle(args: IdleArgs) -> Event<I> { Idle(args) }
-    fn input(args: I) -> Event<I> { Input(args) }
 }
