+++ src/input/src/render.rs
@@ -1,6 +1,33 @@
 use std::any::Any;
+use viewport::Viewport;
 
-use { GenericEvent, RenderArgs, RENDER };
+use { GenericEvent, RENDER };
+
+/// Render arguments
+#[derive(Copy, Clone, PartialEq, Debug)]
+pub struct RenderArgs {
+    /// Extrapolated time in seconds, used to do smooth animation.
+    pub ext_dt: f64,
+    /// The width of rendered area in points.
+    pub width: u32,
+    /// The height of rendered area in points.
+    pub height: u32,
+    /// The width of rendered area in pixels.
+    pub draw_width: u32,
+    /// The height of rendered area in pixels.
+    pub draw_height: u32,
+}
+
+impl RenderArgs {
+    /// Returns viewport information filling entire render area.
+    pub fn viewport(&self) -> Viewport {
+        Viewport {
+            rect: [0, 0, self.draw_width as i32, self.draw_height as i32],
+            window_size: [self.width, self.height],
+            draw_size: [self.draw_width, self.draw_height],
+        }
+    }
+}
 
 /// When the next frame should be rendered
 pub trait RenderEvent {
