+++ src/window/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 
 name = "pistoncore-window"
-version = "0.3.0"
+version = "0.4.0"
 authors = [
     "bvssvni <bvssvni@gmail.com>",
     "Coeuvre <coeuvre@gmail.com>"
@@ -19,7 +19,7 @@ path = "src/lib.rs"
 
 [dependencies.pistoncore-input]
 path = "../input"
-version = "0.2.0"
+version = "0.3.0"
 
 [dependencies]
 libc = "0.1"
